Move log-insights/jsonl-reader.ts to /common/jsonl-reader.ts
