+++ extensions/ql-vscode/src/common/jsonl-reader.ts
